backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
pytesseract
Pillow
paddleocr
paddlepaddle
easyocr